+++ src/bika/lims/catalog/analysis_catalog.py
@@ -36,7 +36,6 @@ CATALOG_ANALYSIS_LISTING = "bika_analysis_catalog"
 # Defining the indexes for this catalog
 _indexes_dict = {
     "sortable_title": "FieldIndex",
-    "getParentUID": "FieldIndex",
     "getDueDate": "DateIndex",
     "getDateReceived": "DateIndex",
     "getResultCaptureDate": "DateIndex",
@@ -50,17 +49,13 @@ _indexes_dict = {
     "getCategoryUID": "FieldIndex",
     "getPointOfCapture": "FieldIndex",
     "getSampleTypeUID": "FieldIndex",
-    # TODO Index "getSamplePointUID" is only used in reports/selection_macros
-    "getSamplePointUID": "FieldIndex",
     "getReferenceAnalysesGroupID": "FieldIndex",
-    "getMethodUID": "FieldIndex",
     "getInstrumentUID": "FieldIndex",
     "getWorksheetUID": "FieldIndex",
     "getOriginalReflexedAnalysisUID": "FieldIndex",
     "getPrioritySortkey": "FieldIndex",
     "getAncestorsUIDs": "KeywordIndex",
     "isSampleReceived": "BooleanIndex",
-    "isRetest": "BooleanIndex",
 }
 # Defining the columns for this catalog
 _columns_list = [
@@ -69,13 +64,10 @@ _columns_list = [
     "getResultCaptureDate",
     "getParentURL",
     "getRequestURL",
-    "getParentTitle",
-    "getParentUID",
     "getClientTitle",
     "getClientURL",
     "getRequestTitle",
     "getResult",
-    "getCalculationUID",
     "getUnit",
     "getKeyword",
     "getCategoryTitle",
@@ -86,12 +78,9 @@ _columns_list = [
     "getReferenceResults",
     # Used in duplicated analysis objects
     "getAnalysisPortalType",
-    "isInstrumentValid",
     # Columns from method
-    "getMethodUID",
     "getMethodTitle",
     "getMethodURL",
-    "getAllowedMethodUIDs",
     "getAnalyst",
     "getAnalystName",
     "getNumberOfRequiredVerifications",
@@ -106,9 +95,6 @@ _columns_list = [
     # defined as a service metacolumn instead of an analysis one
     "getResultOptions",
     "getServiceUID",
-    "getInstrumentEntryOfResults",
-    "getAllowedInstrumentUIDs",
-    "getInstrumentUID",
     "getSampleTypeUID",
     "getClientOrderNumber",
     "getDateReceived",