+++ src/senaite/core/upgrade/v02_00_000.py
@@ -79,12 +79,30 @@ INDEXES_TO_ADD = [
     (SETUP_CATALOG, "department_id", "KeywordIndex"),
 ]
 
+INDEXES_TO_REMOVE = [
+    # List of tuples (catalog_name, index_name)
+    (CATALOG_ANALYSIS_LISTING, "getMethodUID"),
+    (CATALOG_ANALYSIS_LISTING, "isRetest"),
+    (CATALOG_ANALYSIS_LISTING, "getSamplePointUID"),
+    (CATALOG_ANALYSIS_LISTING, "getParentUID"),
+]
+
 METADATA_TO_REMOVE = [
     # Only used in Analyses listing and it's behavior is bizarre, probably
     # because is a dict and requires special care with ZODB
     (CATALOG_ANALYSIS_LISTING, "getInterimFields"),
     # No longer used, see https://github.com/senaite/senaite.core/pull/1709/
-    (CATALOG_ANALYSIS_LISTING, "getAttachmentUIDs")
+    (CATALOG_ANALYSIS_LISTING, "getAttachmentUIDs"),
+    (CATALOG_ANALYSIS_LISTING, "getInstrumentEntryOfResults"),
+    (CATALOG_ANALYSIS_LISTING, "getCalculationUID"),
+    (CATALOG_ANALYSIS_LISTING, "getAllowedInstrumentUIDs"),
+    (CATALOG_ANALYSIS_LISTING, "getMethodUID"),
+    (CATALOG_ANALYSIS_LISTING, "getAllowedMethodUIDs"),
+    (CATALOG_ANALYSIS_LISTING, "getInstrumentUID"),
+    (CATALOG_ANALYSIS_REQUEST_LISTING, "getSamplePointUID"),
+    (CATALOG_ANALYSIS_LISTING, "getParentUID"),
+    (CATALOG_ANALYSIS_LISTING, "getParentTitle"),
+    (CATALOG_ANALYSIS_LISTING, "isInstrumentValid"),
 ]
 
 STALE_WORKFLOW_DEFINITIONS = [
@@ -176,6 +194,9 @@ def upgrade(tool):
     remove_all_supplyorders(portal)
     remove_supplyorder_typeinfo(portal)
     remove_supplyorder_workflow(portal)
+
+    # Remove stale indexes
+    remove_stale_indexes(portal)
 
     # Remove stale metadata
     remove_stale_metadata(portal)
@@ -495,6 +516,24 @@ def del_metadata(catalog_id, column):
                     .format(column, catalog_id))
         return
     catalog.delColumn(column)
+
+
+def remove_stale_indexes(portal):
+    logger.info("Removing stale indexes ...")
+    for catalog_id, index_name in INDEXES_TO_REMOVE:
+        del_index(portal, catalog_id, index_name)
+    logger.info("Removing stale indexes ... [DONE]")
+
+
+def del_index(portal, catalog_id, index_name):
+    logger.info("Removing '{}' index from '{}' ..."
+                .format(index_name, catalog_id))
+    catalog = api.get_tool(catalog_id)
+    if index_name not in catalog.indexes():
+        logger.info("Index '{}' not in catalog '{}' [SKIP]"
+                    .format(index_name, catalog_id))
+        return
+    catalog.delIndex(index_name)
 
 
 def remove_at_portal_types(portal):