+++ CHANGES.rst
@@ -5,6 +5,7 @@ Changelog
 ---------------------
 
 - #1726 Content rules control panel templates styled
+- #1724 Purge stale metadata and indexes from analysis catalog
 - #1720 Fix UnicodeDecodeError for Instrument Import Log View
 - #1719 Service/Method/Calculation relationship handling
 - #1717 Port workflow definitions to senaite namespace